Fix remaining SonarCloud findings
