+++ shifter/engine/provisioner/cloud/gcp/network.py
@@ -3,8 +3,8 @@
 from __future__ import annotations
 
 import logging
-from collections.abc import Callable
-from typing import Any as ExternalValue
+from collections.abc import Callable, Iterable
+from typing import Protocol, cast
 
 from cloud.exceptions import CloudNetworkInventoryError
 from cloud.gcp.base import get_project_id, get_region, import_google_module
@@ -13,13 +13,20 @@ from config import is_gce_range_cell_backend, load_gdc_network_access_config
 logger = logging.getLogger(__name__)
 _SUBNET_CIDR_ANNOTATION = "shifter.dev/subnet-cidr"
 _MANAGED_BY_LABEL = "shifter-provisioner"
-InventoryItem = dict[str, ExternalValue]
+InventoryItem = dict[str, object]
+
+
+class _SubnetworksClient(Protocol):
+    """Subset of the Compute subnetworks client used by inventory."""
+
+    def list(self, *, project: str, region: str) -> Iterable[object]:
+        """Return Compute subnetworks for a project and region."""
 
 
 class GCPNetworkInventory:
     """GCP network inventory implementation of NetworkInventory."""
 
-    def __init__(self, *, gce_subnetworks_client_factory: Callable[[], ExternalValue] | None = None) -> None:
+    def __init__(self, *, gce_subnetworks_client_factory: Callable[[], _SubnetworksClient] | None = None) -> None:
         self._gce_subnetworks_client_factory = gce_subnetworks_client_factory
 
     def list_subnet_cidrs(self, network_id: str) -> list[str]:
@@ -60,7 +67,7 @@ class GCPNetworkInventory:
                 cidrs.append(str(cidr))
         return cidrs
 
-    def _build_gce_subnetworks_client(self) -> ExternalValue:
+    def _build_gce_subnetworks_client(self) -> _SubnetworksClient:
         """Build or reuse the Compute subnetworks client."""
         if self._gce_subnetworks_client_factory:
             return self._gce_subnetworks_client_factory()
@@ -105,13 +112,16 @@ class GCPNetworkInventory:
     @staticmethod
     def _is_managed_gdc_network(item: InventoryItem) -> bool:
         """Return whether a Kubernetes Network belongs to the range plane."""
-        labels = item.get("metadata", {}).get("labels", {}) or {}
+        metadata = item.get("metadata", {})
+        labels = cast(InventoryItem, metadata).get("labels", {}) if isinstance(metadata, dict) else {}
+        if not isinstance(labels, dict):
+            return False
         if labels.get("app.kubernetes.io/managed-by") == "shifter-provisioner":
             return True
         return labels.get("shifter.dev/range-plane") == "gdc-vmruntime"
 
     @staticmethod
-    def _get_field(item: ExternalValue, *names: str) -> ExternalValue:
+    def _get_field(item: object, *names: str) -> object:
         """Read a field from a dict or SDK object."""
         for name in names:
             value = item.get(name) if isinstance(item, dict) else getattr(item, name, None)
@@ -120,15 +130,15 @@ class GCPNetworkInventory:
         return ""
 
     @classmethod
-    def _is_managed_gce_subnetwork(cls, item: ExternalValue) -> bool:
+    def _is_managed_gce_subnetwork(cls, item: object) -> bool:
         """Return whether a Compute subnetwork belongs to the range plane."""
         labels = cls._get_field(item, "labels") or {}
         if not isinstance(labels, dict):
-            labels = dict(labels)
+            return False
         return labels.get("managed-by") == _MANAGED_BY_LABEL
 
     @classmethod
-    def _matches_requested_network(cls, item: ExternalValue, network_id: str) -> bool:
+    def _matches_requested_network(cls, item: object, network_id: str) -> bool:
         """Return whether a Compute subnetwork belongs to the requested range network."""
         if not network_id or network_id.startswith("gcp-range-cells:"):
             return True
@@ -139,13 +149,22 @@ class GCPNetworkInventory:
     def _managed_network_cidrs(item: InventoryItem) -> list[str]:
         """Return CIDRs from the current annotation or legacy route fields."""
         metadata = item.get("metadata", {})
-        annotations = metadata.get("annotations", {}) or {}
+        metadata_dict = cast(InventoryItem, metadata) if isinstance(metadata, dict) else {}
+        annotations = metadata_dict.get("annotations", {}) or {}
+        if not isinstance(annotations, dict):
+            annotations = {}
         annotated_cidr = str(annotations.get(_SUBNET_CIDR_ANNOTATION, "")).strip()
         if annotated_cidr:
             return [annotated_cidr]
 
         legacy_cidrs: list[str] = []
-        for route in item.get("spec", {}).get("routes", []):
+        spec = item.get("spec", {})
+        spec_dict = cast(InventoryItem, spec) if isinstance(spec, dict) else {}
+        routes = spec_dict.get("routes", [])
+        routes_list = routes if isinstance(routes, list) else []
+        for route in routes_list:
+            if not isinstance(route, dict):
+                continue
             cidr = str(route.get("to", "")).strip()
             if cidr:
                 legacy_cidrs.append(cidr)
